Add main movility to the objects in the game
The hens are created with an specific direction and already get into the
barn.
All the barns are located in the scenario and have track of how many hens
have arrived.
When pushing the  the B button of the control, the number of hens in
barns is got.

Implementation details:
A scenario and scenarioObject class was created to encapsulate the
behavior of all scenario objects.
Barn, Hen and JumpingHen classes are subclasses of ScenarioObject class.
Hen and JumpingHen classes implement move() method to control their
behavior in the scenario.
In Game.cpp class movement and score is managed on in APP_LOOP event.

Signed-off-by: Nelson Victor Cruz Hernandez <victor.cruz.isc@gmail.com>